test(e2e): Move utils to a support folder
